refactor: adota frontend, backend e shared para a feature commons
