--- src/features/common/formatters/cpf.ts
@@ -1,2 +1,0 @@
-export const cpfFormatter = (value: string): string =>
-    value.replace(/(\d{3})(\d{3})(\d{3})(\d{2})/, "$1.$2.$3-$4");
