--- src/features/common/frontend/WithSearch/context.ts
@@ -1,10 +1,0 @@
-"use client";
-import { createContext } from "react";
-
-export type SearchModel = {
-    searchString: string;
-};
-
-export const SearchContext = createContext<SearchModel>({
-    searchString: "",
-});
